chore: refactor coverage types to avoid mixing node and runtime types
